requirements.txt
django==4.0.0
djangorestframework
psycopg2-binary
pytest
pytest-django
factory_boy
pytest_factoryboy
